Fix componente title nad board
