+++ src/body/board/Board.tsx
@@ -1,10 +1,11 @@
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import more from "../../assets/more.svg";
 import styles from "./styles.module.css";
 import { useContext, useEffect, useState } from "react";
-import { fetchDelete, fetchGet, fetchPatch } from "../../utils/functions-fetch";
+import { fetchGet } from "../../utils/functions-fetch";
 import { URL } from "../../utils/variables";
 import { Page } from "../../App/App";
+import Title from "./title/Title";
 
 const board = {
   boardData: {},
@@ -14,11 +15,6 @@ const board = {
 
 function Board() {
   const [dataBoard, setDataBoard] = useState(board);
-  const [dropDown, setDropDown] = useState(false);
-  const [editData, setEditData] = useState(false);
-  const [titleEdit, setTitleEdit] = useState("");
-
-  const navigate = useNavigate();
 
   const { id } = useParams();
 
@@ -30,46 +26,6 @@ function Board() {
 
   const { currentPage, setCurrentPage } = pageContext;
   //
-
-  const onSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const response = await fetchPatch(URL, `/boards/${id}`, {
-      title: titleEdit,
-    });
-    console.log(response.data.data);
-    setEditData(!editData);
-    setCurrentPage(!currentPage);
-  };
-
-  const inputEdit = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = event.target.value;
-    setTitleEdit(newValue);
-  };
-
-  const optionMore = () => {
-    setDropDown(!dropDown);
-  };
-
-  const optionDropDown = async (
-    event: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
-    const option = event.currentTarget.value;
-    setDropDown(false);
-    if (option === "Edit") {
-      setEditData(!editData);
-    } else if (option === "Delete") {
-      const response = await fetchDelete(URL, `/boards/${id}`, {
-        headers: {
-          id: user.id,
-          username: user.username,
-          authorization: `Bearer ${user.token}`,
-        },
-      });
-      navigate("/");
-      console.log(response);
-    }
-  };
 
   const object = localStorage.getItem("user");
   const user = object && JSON.parse(object);
@@ -96,42 +52,21 @@ function Board() {
       className={styles.bodyBoard}
       style={{ background: `${dataBoard.color}` }}>
       <div className={styles.containerBoard}>
-        <div className={styles.containerTitle}>
-          <form onSubmit={onSubmit}>
-            <input
-              type="text"
-              className={styles.inputDisplay}
-              style={{ display: editData ? "flex" : "none" }}
-              onChange={inputEdit}
-            />
-            <button style={{ display: "none" }}></button>
-          </form>
-          <h1
-            className={styles.title}
-            style={{ display: editData ? "none" : "block" }}>
-            {dataBoard.title}
-          </h1>
-          <div className={styles.containerDropDown}>
-            <img onClick={optionMore} src={more} alt="more" />
-            <div
-              className={styles.dropDown}
-              style={{ display: dropDown ? "flex" : "none" }}>
-              <button
-                className={styles.option}
-                value="Edit"
-                onClick={optionDropDown}>
-                Edit
-              </button>
-              <button
-                className={styles.option}
-                value="Delete"
-                onClick={optionDropDown}>
-                Delete
-              </button>
-            </div>
-          </div>
-        </div>
+        <Title
+          dataBoard={dataBoard}
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
+        />
         <div className={styles.containerMenuCards}>
+          <form className={styles.formCreateTable}>
+            <div className={styles.containerInputTable}>
+              <p className={styles.titleForm}>List Title</p>
+              <input className={styles.inputForm} type="text" />
+            </div>
+            <button className={styles.buttonCreateTable}>
+              Create new list
+            </button>
+          </form>
           <div className={styles.containerTasks}>
             <div className={styles.containerFlex}>
               <div>
