+++ src/body/board/title/styles.module.css
@@ -1,0 +1,56 @@
+.containerTitle {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
+
+.title {
+  color: #171717;
+  font-family: Inter;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 32px;
+}
+
+.containerDropDown {
+  position: relative;
+}
+
+.dropDown {
+  position: absolute;
+  display: flex;
+  min-width: 96px;
+  padding: 8px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 4px;
+  border-radius: 6px;
+  background: #fff;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.25);
+}
+
+.option {
+  background-color: white;
+  border: none;
+
+  color: #171717;
+  font-family: Inter;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 20px;
+}
+
+.inputDisplay {
+  display: flex;
+  height: 35px;
+  padding: 8px 12px;
+  align-items: flex-start;
+  gap: 4px;
+  align-self: stretch;
+  border-radius: 6px;
+  border: 1px solid #d4d4d4;
+  background: #fff;
+  box-sizing: border-box;
+}
